# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/asset/maintenance": {
            "get": {
                "description": "Returns a paginated list of assets maintenance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get All Asset Maintenance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetAllAssetMaintenance"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an asset maintenance by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Post Create Asset Maintenance By Id",
                "parameters": [
                    {
                        "description": "Create Asset Maintenance Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RequestCreateUpdateAssetMaintenance"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseCreateAssetMaintenance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/asset/maintenance/schedule": {
            "get": {
                "description": "Returns a list of assets maintenance schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get All Asset Maintenance Schedule",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetAllAssetMaintenanceSchedule"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/asset/maintenance/{id}": {
            "put": {
                "description": "Create an asset maintenance by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Put Update Asset Maintenance By Id",
                "parameters": [
                    {
                        "description": "Put Update Asset Maintenance Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RequestCreateUpdateAssetMaintenance"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Id of asset maintenance",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponsePutUpdateAssetMaintenance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanentally delete asset maintenance by id",
                "tags": [
                    "Asset"
                ],
                "summary": "Delete Asset Maintenance By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of asset maintenance",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseDeleteAssetMaintenanceById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/asset/most-context/{targetCol}": {
            "get": {
                "description": "Returns a list of most appear item in asset maintenance by given field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get Most Context Asset Maintenance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target Column to Analyze (such as: maintenance_day)",
                        "name": "targetCol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetMostContext"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.AssetMaintenance": {
            "type": "object",
            "properties": {
                "asset_placement_id": {
                    "description": "FK - Asset Placement",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "description": "FK - Admin",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "maintenance_by": {
                    "description": "FK - Technician",
                    "type": "string"
                },
                "maintenance_day": {
                    "type": "string"
                },
                "maintenance_hour_end": {
                    "type": "string"
                },
                "maintenance_hour_start": {
                    "type": "string"
                },
                "maintenance_notes": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.AssetMaintenanceSchedule": {
            "type": "object",
            "properties": {
                "asset_category": {
                    "type": "string"
                },
                "asset_name": {
                    "description": "FK - Asset",
                    "type": "string"
                },
                "asset_qty": {
                    "description": "FK - Asset Placement",
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "maintenance_day": {
                    "type": "string"
                },
                "maintenance_hour_end": {
                    "type": "string"
                },
                "maintenance_hour_start": {
                    "type": "string"
                },
                "maintenance_notes": {
                    "type": "string"
                },
                "telegram_is_valid": {
                    "type": "boolean"
                },
                "telegram_user_id": {
                    "type": "string"
                },
                "username": {
                    "description": "FK - Technician",
                    "type": "string"
                }
            }
        },
...
        "entity.Metadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "entity.RequestCreateUpdateAssetMaintenance": {
            "type": "object",
            "required": [
                "asset_placement_id",
                "maintenance_by",
                "maintenance_day",
                "maintenance_hour_end",
                "maintenance_hour_start"
            ],
            "properties": {
                "asset_placement_id": {
                    "type": "string"
                },
                "maintenance_by": {
                    "type": "string"
                },
                "maintenance_day": {
                    "type": "string"
                },
                "maintenance_hour_end": {
                    "type": "string"
                },
                "maintenance_hour_start": {
                    "type": "string"
                },
                "maintenance_notes": {
                    "type": "string"
                }
            }
        },
...
        "entity.ResponseCreateAssetMaintenance": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset maintenance created successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseDeleteAssetMaintenanceById": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset maintenance deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetAllAssetMaintenance": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.AssetMaintenance"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset maintenance fetched"
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Metadata"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "entity.ResponseGetAllAssetMaintenanceSchedule": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.AssetMaintenanceSchedule"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset maintenance schedule fetched"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetMostContext": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StatsContextTotal"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset fetched"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponsePutUpdateAssetMaintenance": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset maintenance updated successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.StatsContextTotal": {
            "type": "object",
            "properties": {
                "context": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}